:sparkles: Add supported swift type
